Component rendering using an inlined version of markaby
I think the published gem is not the latest version (the version number of 0.9.0 has not changed although the code has).

So I've inlined the latest version of Markaby so the support for data attributes can be included.

The component now renders with a reference to the Stimul8 Stimulus controller